+++ src/test/python/apache/aurora/client/cli/test_inspect.py
@@ -142,7 +142,9 @@ Process 'process':
             "http": {
                 "graceful_shutdown_endpoint": "/quitquitquit",
                 "port": "health",
-                "shutdown_endpoint": "/abortabortabort"}},
+                "shutdown_endpoint": "/abortabortabort",
+                "graceful_shutdown_wait_secs": 5,
+                "shutdown_wait_secs": 5}},
         "priority": 0}
 
     mock_output = "\n".join(mock_stdout)