+++ src/test/python/apache/aurora/executor/bin/test_thermos_executor_entry_point.py
@@ -34,6 +34,7 @@ class ThermosExecutorMainTest(unittest.TestCase):
     mock_options.execute_as_user = False
     mock_options.nosetuid = False
     mock_options.announcer_ensemble = None
+    mock_options.stop_timeout_in_secs = 1
     with patch(
         'apache.aurora.executor.bin.thermos_executor_main.dump_runner_pex',
         return_value=mock_dump_runner_pex):